refactor(terminal): extract pane runtime helpers
